Added comments and did a little cleanup.
