+++ snapshot/bundlestore/server_test.go
@@ -39,6 +39,8 @@ func (f *FakeStore) Write(name string, data io.Reader) error {
 	}
 }
 
+//TODO: an end-end test that uses real a real store and real bundles.
+
 func TestServer(t *testing.T) {
 	// Construct server with a fake store and random port address.
 	store := &FakeStore{files: map[string][]byte{}}
@@ -46,7 +48,7 @@ func TestServer(t *testing.T) {
 	listener.Close()
 	addr := listener.Addr().String()
 	go MakeServer(store, Addr(addr)).Serve()
-	time.Sleep(50 * time.Millisecond) // needed for travis-ci?
+	//time.Sleep(50 * time.Millisecond) // needed for travis-ci?
 
 	rootUri := "http://" + addr + "/bundle/"
 	client := &http.Client{Timeout: 1 * time.Second}
